Refactor: buttons in Login, MainPage
