+++ src/axios/requestIntercepters.ts
@@ -1,11 +1,19 @@
+import { InternalAxiosRequestConfig } from "axios";
 import { STORAGE_KEY } from "@/constants";
 import axiosInstance from "./instance";
 
 axiosInstance.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem(STORAGE_KEY.TOKEN);
-    config.headers.Authorization = !!token ? `Bearer ${token}` : "";
-    return config;
-  },
+  (config) => handleRequestIntercept(config),
   (error) => Promise.reject(error),
 );
+
+const handleRequestIntercept = (config: InternalAxiosRequestConfig) => {
+  const token = localStorage.getItem(STORAGE_KEY.TOKEN);
+
+  if (config?.headers.includeAuthorization && !!token) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+
+  delete config.headers.includeAuthorization;
+  return config;
+};
