+++ src/utils/httpRequest.ts
@@ -1,38 +1,9 @@
-import axios, { InternalAxiosRequestConfig } from "axios";
-import { STORAGE_KEY, URL } from "@/constants";
+import axiosInstance from "@/axios";
+import { URL } from "@/constants";
 import { RequestParams } from "@/interfaces/httpRequest";
 
-interface HandleError {
-  [status: number]: () => void;
-}
-
-const instance = axios.create({
-  baseURL: "",
-  headers: {
-    common: {
-      "Content-Type": "application/json",
-    },
-  },
-});
-
-instance.interceptors.request.use(
-  (config) => handleRequestIntercept(config),
-  (error) => Promise.reject(error),
-);
-
-const handleRequestIntercept = (config: InternalAxiosRequestConfig) => {
-  const token = localStorage.getItem(STORAGE_KEY.TOKEN);
-
-  if (config?.headers.includeAuthorization && !!token) {
-    config.headers.Authorization = `Bearer ${token}`;
-  }
-
-  delete config.headers.includeAuthorization;
-  return config;
-};
-
 const fetchRequest = <TQueryParams>({ path, method, params, data, isMock, config }: RequestParams<TQueryParams>) => {
-  return instance(`${isMock ? "" : URL.API}${path}`, {
+  return axiosInstance(`${isMock ? "" : URL.API}${path}`, {
     headers: config,
     method,
     data: JSON.stringify(data),
